+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ltayra-y <ltayra-y@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/19 16:01:56 by ltayra-y          #+#    #+#             */
-/*   Updated: 2024/11/20 14:28:53 by ltayra-y         ###   ########.fr       */
+/*   Updated: 2024/11/21 14:20:35 by ltayra-y         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <stdarg.h>
 # include <unistd.h>
 
-int		nbrlen(int n);
+size_t	nbrlen(int n);
 int		print_char(char c);
 void	ft_putchar(char c);
 void	ft_putstr(char *s);
@@ -24,7 +24,7 @@ size_t	ft_strlen(const char *s);
 int		print_str(char *s);
 void	ft_putnbr(int n);
 int		print_nbr(int n);
-int		nbrlen_un(unsigned int n);
+size_t	nbrlen_un(unsigned int n);
 void	putnbr_un(unsigned int n);
 int		print_unsigned(unsigned int n);
 int		print_ptr(unsigned long ptr);
